+++ docs/whats_new.rst
@@ -9,6 +9,10 @@ This information applies to the
 Version 0.10
 ------------
 
+- 2023-08-18: Upgrade to Blender LTS version 3.6 (Python 3.10) and VTK 9.2.6.
+  The previous Blender LTS version 3.3 (Python 3.10) still works as well
+  with VTK 9.2.6. This upgrade required no class changes.
+
 - 2023-06-09: Fixed a bug where negative scalar values were colored incorrectly
   with Color Mapper node (absolute values of data was shown).